plan, executor: remove AdminStmt, AlterTableStmt, AnalyzeTableStmt, DropStatsStmt, CreateIndexStmt, CreateTableStmt, DoStmt, DropTableStmt, DropIndexStmt, RenameTableStmt, SetStmt, ShowStmt from resolver.go
