+++ plan/planbuilder.go
@@ -219,6 +219,10 @@ func (b *planBuilder) buildSet(v *ast.SetStmt) Plan {
 			IsSystem: vars.IsSystem,
 		}
 		if _, ok := vars.Value.(*ast.DefaultExpr); !ok {
+			if cn, ok2 := vars.Value.(*ast.ColumnNameExpr); ok2 && cn.Name.Table.L == "" {
+				// Convert column name expression to string value expression.
+				vars.Value = ast.NewValueExpr(cn.Name.Name.O)
+			}
 			mockTablePlan := TableDual{}.init(b.allocator, b.ctx)
 			mockTablePlan.SetSchema(expression.NewSchema())
 			assign.Expr, _, b.err = b.rewrite(vars.Value, mockTablePlan, nil, true)
@@ -532,7 +536,7 @@ func (b *planBuilder) buildShow(show *ast.ShowStmt) Plan {
 		User:   show.User,
 	}.init(b.allocator, b.ctx)
 	resultPlan = p
-	switch show.Tp {
+	switch showTp := show.Tp; showTp {
 	case ast.ShowProcedureStatus:
 		p.SetSchema(buildShowProcedureSchema())
 	case ast.ShowTriggers:
@@ -542,6 +546,13 @@ func (b *planBuilder) buildShow(show *ast.ShowStmt) Plan {
 	case ast.ShowWarnings:
 		p.SetSchema(buildShowWarningsSchema())
 	default:
+		switch showTp {
+		case ast.ShowTables, ast.ShowTableStatus:
+			if p.DBName == "" {
+				b.err = ErrNoDB
+				return nil
+			}
+		}
 		p.SetSchema(buildShowSchema(show))
 	}
 	for i, col := range p.schema.Columns {
@@ -549,6 +560,9 @@ func (b *planBuilder) buildShow(show *ast.ShowStmt) Plan {
 	}
 	var conditions []expression.Expression
 	if show.Pattern != nil {
+		show.Pattern.Expr = &ast.ColumnNameExpr{
+			Name: &ast.ColumnName{Name: p.Schema().Columns[0].ColName},
+		}
 		expr, _, err := b.rewrite(show.Pattern, p, nil, false)
 		if err != nil {
 			b.err = errors.Trace(err)
